Entity

+++ src/HrHub.Domain/Entities/SqlDbEntities/ExamQuestion.cs
@@ -1,15 +1,15 @@
 ﻿using HrHub.Core.Domain.Entity;
 using System.ComponentModel.DataAnnotations.Schema;
 namespace HrHub.Domain.Entities.SqlDbEntities;
-public class ExamQuestion : TypeCardEntity<int>
+public class ExamQuestion : TypeCardEntity<long>
 {
     public ExamQuestion()
     {
         ExamAnswers = new HashSet<ExamAnswer>();
     }
     [ForeignKey("ExamTopicId")]
-    public int ExamTopicId { get; set; }
-    public int Question { get; set; }
+    public long ExamTopicId { get; set; }
+    public long Question { get; set; }
 
     public virtual ICollection<ExamAnswer> ExamAnswers { get; set; }
     public virtual ExamTopic ExamTopics { get; set; }
