+++ src/HrHub.Domain/Entities/SqlDbEntities/Language.cs
@@ -2,7 +2,7 @@
 
 namespace HrHub.Domain.Entities.SqlDbEntities
 {
-    public class Language : TypeCardEntity<int>
+    public class Language : TypeCardEntity<long>
     {
         public string LangCode { get; set; }
     }
