+++ src/HrHub.Domain/Entities/SqlDbEntities/TrainingLevel.cs
@@ -2,9 +2,9 @@
 
 namespace HrHub.Domain.Entities.SqlDbEntities
 {
-    public class TrainingLevel : TypeCardEntity<int>
+    public class TrainingLevel : TypeCardEntity<long>
     {
-        public int Priority { get; set; }
+        public long Priority { get; set; }
         public string Color { get; set; }
     }
 }
